Refactored all tests to work with new HandlaData class
