--- CountPostIts.ConsoleUI.Tests/ImageRecognitionLibraryTests.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using NUnit.Framework;
-
-namespace CountPostIts.ConsoleUI.Tests
-{
-    public class ImageRecognitionLibraryTests
-    {
-        [SetUp]
-        public void SetUp()
-        {
-        }
-
-        [Test]
-        public void test()
-        {
-
-        }
-    }
-}
