+++ CountPostIts.ConsoleUI/CountPostIts.ConsoleUI.csproj
@@ -62,12 +62,12 @@
     <Compile Include="IFileWrapper.cs" />
     <Compile Include="IInformationWrapper.cs" />
     <Compile Include="InformationWrapper.cs" />
-    <Compile Include="IReturnResults.cs" />
+    <Compile Include="IHandleData.cs" />
     <Compile Include="Options.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Config.cs" />
-    <Compile Include="ReturnResults.cs" />
+    <Compile Include="HandleData.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
