+++ CountPostIts.ConsoleUI/IInformationWrapper.cs
@@ -9,7 +9,5 @@ namespace CountPostIts.ConsoleUI
     public interface  IInformationWrapper
     {
         int CallCountPostits(string filename);
-
-        bool HasPostits(string filename);
     }
 }
