+++ CountPostIts.ConsoleUI/Program.cs
@@ -14,7 +14,9 @@ namespace CountPostIts.ConsoleUI
         static void Main(string[] args)
         {
             // add parsing library
-            var config = new Config(new FileWrapper(), new InformationWrapper());
+            var config = new Config(
+                new FileWrapper(), new HandleData(new InformationWrapper())
+                );
             // add display results
         }
     }
